assemblyscript compiler back into worker
