--- wasmaudioworklet/synth1/assemblyscriptcompiler.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-
-    </head>
-    <body>
-        <script src="https://cdn.jsdelivr.net/gh/AssemblyScript/binaryen.js@89.0.0-nightly.20191113/index.js"></script>
-        <script src="https://cdn.jsdelivr.net/gh/AssemblyScript/AssemblyScript@0.8.1/dist/assemblyscript.js"></script>
-        <script src="https://cdn.jsdelivr.net/gh/AssemblyScript/AssemblyScript@0.8.1/dist/asc.js"></script>
-        <script>
-            async function compileAssemblyScript(sources, options, entrypoint) {
-                if (typeof sources === "string") sources = { "input.ts": sources };
-                const output = Object.create({
-                    stdout: asc.createMemoryStream(),
-                    stderr: asc.createMemoryStream()
-                });
-                var argv = [
-                    "--binaryFile", "binary",
-                    "--textFile", "text",
-                ];
-                Object.keys(options || {}).forEach(key => {
-                    var val = options[key];
-                    if (Array.isArray(val)) val.forEach(val => argv.push("--" + key, String(val)));
-                    else argv.push("--" + key, String(val));
-                });
-                asc.main(entrypoint ? argv.concat(entrypoint) : argv.concat(Object.keys(sources)), {
-                    stdout: output.stdout,
-                    stderr: output.stderr,
-                    readFile: name => sources.hasOwnProperty(name) ? sources[name] : null,
-                    writeFile: (name, contents) => output[name] = contents,
-                    listFiles: () => []
-                });
-                
-                return output;
-            }
-
-            async function compile() {
-                const sources = await fetch('wasmsynthassemblyscriptsources.json').then(r => r.json());
-                
-                const {stderr, text, binary} = compileAssemblyScript(sources,
-                        { "runtime": "none", "optimizeLevel": 3, "shrinkLevel": 1},
-                        'index.ts');
-                
-                console.log(stderr);
-            }
-            compile();
-        </script>        
-    </body>
-</html>
