--- wasmaudioworklet/synth1/package.json
@@ -1,15 +1,0 @@
-{
-  "scripts": {
-    "asbuild": "asc --runtime stub assembly/index.ts -o build/index.wasm -Oz",
-    "fastbuild": "asc --runtime stub assembly/index.ts -o build/index.wasm -t build/index.wat",
-    "createbrowsersourcebundle": "node createbrowsertsbundle.js",
-    "test": "asp --verbose",
-    "test:ci": "asp --summary"
-  },
-  "dependencies": {
-    "assemblyscript": "^0.27.14"
-  },
-  "devDependencies": {
-    "@as-pect/cli": "^8.1.0"
-  }
-}
